Use Long to represent the max number of chars to accept when binding a form from Json

+++ web/play-java-forms/src/main/java/play/data/DynamicForm.java
@@ -284,7 +284,7 @@ public class DynamicForm extends Form<DynamicForm.Dynamic> {
 
   @Override
   public DynamicForm bind(
-      Lang lang, TypedMap attrs, JsonNode data, int maxChars, String... allowedFields) {
+      Lang lang, TypedMap attrs, JsonNode data, long maxChars, String... allowedFields) {
     return bind(
         lang,
         attrs,
